requirements.txt
Flask
BeautifulSoup4
pep8
pytest-cov
gunicorn
newrelic
hypothesis
